fix: add service_id validation, token expiry constant, and test cleanup
- Add ACCESS_TOKEN_EXPIRY_SECONDS constant (lost during rebase)
- Add service_id validation to resource create endpoint
- Add service_id field to resource create and detail responses
- Move _get_bearer_token to test base class to avoid duplication
- Update tests to use product_analytics instead of posthog_analytics
